Express nmda synaptic current I_syn_nmda_dp as a function of Imem and Inmda
Inmda is the current through the Vnmda transistor.
Currently the equations are defined such that the Vnmda is an input parameter
and Inmda is computed accordingly.
This helps to visualize the sigmoidal behavior of the Differential Pair
But, to avoid weird behavior of the synaptic current (e.g. to make sure that
the synaptic current will depend also on the absolute value of Vmem, or Imem,
and not only on the delta=Vnmda-Vmem), it is more accurate to express
the output of the differential pair as a function of currents.

Note that this makes also the equations consistent with the idea of expressing
all the equations in terms of currents and not in terms of voltages.

TODO: Consider replacing the input Vnmda with the current directly.